+++ src/errors/index.ts
@@ -43,4 +43,5 @@ export { NotExtendedError } from "./not-extended-error";
 export { NetworkAuthenticationRequiredError } from "./network-authentication-required-error";
 export { BaseHttpError } from "./base-http-error";
 export { UnknownHttpError } from "./unknown-http-error";
-export type { ClientErrors, ServerErrors, TypedFetchError } from "./helpers";
+export { httpErrors } from "./helpers";
+export type { ClientErrors, HttpErrors, ServerErrors, TypedFetchError } from "./helpers";